GPT2-Chest-X-Ray-Report-Generation-master/current_requirements.txt
absl-py==2.0.0
astunparse==1.6.3
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
cachetools==5.3.2
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1700303426725/work/certifi
cftime @ file:///D:/bld/cftime_1656803796458/work
charset-normalizer==3.3.2
click==8.1.7
colorama @ file:///C:/b/abs_a9ozq0l032/croot/colorama_1672387194846/work
cycler==0.11.0
Cython==0.29.14
debugpy @ file:///C:/ci/debugpy_1637091911212/work
decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
efficientnet==1.1.1
entrypoints @ file:///C:/ci/entrypoints_1649926659561/work
filelock==3.12.2
Flask==2.2.5
flatbuffers==23.5.26
fonttools==4.38.0
fsspec==2023.1.0
gast==0.3.3
gensim==3.8.3
google-auth==2.24.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.59.3
h5py==2.10.0
huggingface-hub==0.16.4
idna==3.6
imageio==2.31.2
importlib-metadata==6.7.0
ipykernel @ file:///C:/b/abs_21ykzkm7y_/croots/recipe/ipykernel_1662361803478/work
ipython @ file:///C:/Windows/TEMP/abs_45b5zb1l7q/croots/recipe/ipython_1659529855872/work
itsdangerous==2.1.2
jedi @ file:///C:/ci/jedi_1644297241925/work
Jinja2==3.1.2
joblib==1.3.2
jupyter_client @ file:///C:/b/abs_d8fk_kz9zk/croot/jupyter_client_1676330195659/work
jupyter_core @ file:///C:/b/abs_b8r5lb5eus/croot/jupyter_core_1668084445779/work
keras==2.11.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
libclang==16.0.6
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.5.3
matplotlib-inline @ file:///C:/ci/matplotlib-inline_1661934089728/work
nest-asyncio @ file:///C:/b/abs_3a_4jsjlqu/croot/nest-asyncio_1672387322800/work
netCDF4 @ file:///D:/bld/netcdf4_1656505682593/work
networkx==2.6.3
nlg-eval @ git+https://github.com/Maluuba/nlg-eval.git@8c4c41a617e1ceba1aa73fbf6a6074d7204d897d
nltk==3.8.1
numpy @ file:///D:/bld/numpy_1649806536290/work
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.2
pandas==1.3.5
parso @ file:///opt/conda/conda-bld/parso_1641458642106/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow==9.5.0
pip-check==2.8.1
pipdeptree==2.9.6
prompt-toolkit @ file:///C:/b/abs_6coz5_9f2s/croot/prompt-toolkit_1672387908312/work
protobuf==3.19.6
psutil @ file:///C:/Windows/Temp/abs_b2c2fd7f-9fd5-4756-95ea-8aed74d0039flsd9qufz/croots/recipe/psutil_1656431277748/work
pyasn1==0.5.1
pyasn1-modules==0.3.0
Pygments @ file:///opt/conda/conda-bld/pygments_1644249106324/work
pyparsing==3.1.1
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
pytz==2023.3.post1
PyWavelets==1.3.0
pywin32==305.1
PyYAML==6.0.1
pyzmq @ file:///C:/ci/pyzmq_1657597757589/work
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
safetensors==0.4.1
scikit-image==0.19.3
scikit-learn==1.0.2
scipy==1.4.1
six @ file:///tmp/build/80754af9/six_1644875935023/work
smart-open==6.4.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.3.0
tensorflow-estimator==2.3.0
tensorflow-intel==2.11.1
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
terminaltables==3.1.10
Theano==1.0.5
threadpoolctl==3.1.0
tifffile==2021.11.2
tokenizers==0.13.3
tornado @ file:///C:/ci/tornado_1662476933490/work
tqdm==4.66.1
traitlets @ file:///C:/b/abs_e5m_xjjl94/croot/traitlets_1671143896266/work
transformers==4.30.2
typing_extensions==4.7.1
urllib3==2.0.7
wcwidth @ file:///Users/ktietz/demo/mc3/conda-bld/wcwidth_1629357192024/work
Werkzeug==2.2.3
wincertstore==0.2
wrapt==1.16.0
xdg==6.0.0
zipp==3.15.0
